+++ Runtime/IndexedWidgetBuilder.cs
@@ -1,4 +1,0 @@
-namespace UniMob.UI
-{
-    public delegate Widget IndexedWidgetBuilder(BuildContext context, int index);
-}